Basket Update
update functions add stock contol

+++ src/main/java/com/works/services/OrdersService.java
@@ -3,9 +3,11 @@ package com.works.services;
 import com.works.entities.Basket;
 import com.works.entities.Customer;
 import com.works.entities.Orders;
+import com.works.entities.Product;
 import com.works.repositories.BasketRepository;
 import com.works.repositories.CustomerRepository;
 import com.works.repositories.OrdersRepository;
+import com.works.repositories.ProductRepository;
 import com.works.utils.REnum;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -22,31 +24,32 @@ public class OrdersService {
     final BasketRepository basketRepository;
     final CustomerRepository customerRepository;
     final HttpSession session;
+    final ProductRepository productRepository;
 
-    public OrdersService(OrdersRepository ordersRepository, ProductService productService, BasketRepository basketRepository, CustomerRepository customerRepository, HttpSession session) {
+    public OrdersService(OrdersRepository ordersRepository, ProductService productService, BasketRepository basketRepository, CustomerRepository customerRepository, HttpSession session, ProductRepository productRepository) {
         this.ordersRepository = ordersRepository;
         this.productService = productService;
         this.basketRepository = basketRepository;
         this.customerRepository = customerRepository;
         this.session = session;
+        this.productRepository = productRepository;
     }
 
-    public ResponseEntity add(Orders orders) {
+    public ResponseEntity add() {
         Map<REnum, Object> hm = new LinkedHashMap<>();
+        Orders orders=new Orders();
         int sum=0;
         Customer customer= (Customer) session.getAttribute("customer");
         if(customer!=null) {
-            System.out.println("customer mail"+customer.getEmail());
-            List<Basket> baskets = basketRepository.findByCreatedByEqualsAndStatusFalse(customer.getEmail());
+            List<Basket> baskets = basketRepository.findByCreatedByEqualsAndStatusFalse(customer.getEmail() );
 
             if (baskets.size() > 0) {
                 orders.setCustomer(customer);
                 orders.setBaskets(baskets);
                 for (Basket item : baskets) {
                     sum = sum + item.getProduct().getPrice() * item.getQuantity();
-                    Optional<Basket> optionalBasket = basketRepository.findById(item.getId());
-                    optionalBasket.get().setStatus(true);
-                    basketRepository.saveAndFlush(optionalBasket.get());
+                    item.setStatus(true);
+                    basketRepository.saveAndFlush(item);
                 }
                 orders.setTotal(sum);
                 ordersRepository.save(orders);
@@ -65,12 +68,31 @@ public class OrdersService {
         }
 
     }
+
+
+
     public ResponseEntity delete(long id) {
         Map<REnum, Object> hm = new LinkedHashMap<>();
         try {
-            ordersRepository.deleteById(id);
-            hm.put(REnum.status, true);
-            return new ResponseEntity<>(hm, HttpStatus.OK);
+           Optional<Orders> optionalOrders=ordersRepository.findById(id);
+           if(optionalOrders.isPresent()){
+               Orders orders=optionalOrders.get();
+               ordersRepository.deleteById(id);
+               List<Basket> baskets=orders.getBaskets();
+               for(Basket basket:baskets){
+                   Integer quantity=basket.getQuantity();
+                   Product product=basket.getProduct();
+                   product.setStockQuantity(product.getStockQuantity()+quantity);
+                   productRepository.saveAndFlush(product);
+               }
+                hm.put(REnum.status, true);
+               hm.put(REnum.message, "Order cancellation successful.");
+                return new ResponseEntity<>(hm, HttpStatus.OK);
+           }else{
+               hm.put(REnum.status, false);
+               hm.put(REnum.message, "There is not such order id");
+               return new ResponseEntity<>(hm, HttpStatus.BAD_REQUEST);
+           }
         } catch (Exception ex) {
             hm.put(REnum.status, false);
             hm.put(REnum.error, ex.getMessage());
@@ -81,7 +103,6 @@ public class OrdersService {
 
     public ResponseEntity getOrderByOrder_Id(long id){
         Map<REnum, Object> hm = new LinkedHashMap<>();
-      /*  List<Basket> baskets=  basketRepository.findByCustomer_Orders_IdIs(id);*/
         List<Orders> orders=ordersRepository.findByIdIs(id);
         hm.put(REnum.result,orders);
         return new ResponseEntity<>(hm,HttpStatus.OK);
@@ -90,20 +111,29 @@ public class OrdersService {
 
     public ResponseEntity list() {
         Map<REnum, Object> hm = new HashMap<>();
-        List<Orders> categoryList = ordersRepository.findAll();
+        List<Orders> ordersList = ordersRepository.findAll();
         hm.put(REnum.status, true);
-        hm.put(REnum.result, categoryList);
+        hm.put(REnum.result, ordersList);
         return new ResponseEntity<>(hm, HttpStatus.OK);
 
 
     }
 
+    public ResponseEntity getOrderByCustomer() {
+        Map<REnum, Object> hm = new HashMap<>();
+        try {
+            Customer customer= (Customer) session.getAttribute("customer");
+            List<Orders> orders=ordersRepository.findByCustomer_Id(customer.getId());
+            hm.put(REnum.status, true);
+            hm.put(REnum.result, orders);
+            return new ResponseEntity<>(hm, HttpStatus.OK);
+        }catch (Exception ex){
+            hm.put(REnum.status, false);
+            hm.put(REnum.error, ex.getMessage());
+            return new ResponseEntity<>(hm, HttpStatus.BAD_REQUEST);
+        }
 
 
-   /* public  Integer generateOrderId() {
-        Random random = new Random();
-        int num = random.nextInt(100000000);
-        String formatted = String.format("%05d", num);
-        return num;
-    }*/
+    }
+
 }
